handle clearing textarea
